Allow both types of imports
