+++ src/react/index.js
@@ -1,6 +1,1 @@
-import { Swiper } from './swiper.js';
-import { SwiperSlide } from './swiper-slide.js';
-
-export { useSwiperSlide, useSwiper } from './context.js';
-
-export { Swiper, SwiperSlide };
+export * from './swiper-react.js';
