+++ src/react/swiper-react.js
@@ -1,0 +1,6 @@
+import { Swiper } from './swiper.js';
+import { SwiperSlide } from './swiper-slide.js';
+
+export { useSwiperSlide, useSwiper } from './context.js';
+
+export { Swiper, SwiperSlide };
