+++ src/solid/index.js
@@ -1,4 +1,1 @@
-export { Swiper } from './swiper.js';
-export { SwiperSlide } from './swiper-slide.js';
-
-export { useSwiperSlide, useSwiper } from './context.js';
+export * from './swiper-solid.js';
