+++ src/solid/swiper-solid.js
@@ -1,0 +1,4 @@
+export { Swiper } from './swiper.js';
+export { SwiperSlide } from './swiper-slide.js';
+
+export { useSwiperSlide, useSwiper } from './context.js';
